Refactor: Rename openapi and schema packages to openapigen and schemagen
