+++ specification/openapigen/openapi.go
@@ -1,4 +1,4 @@
-package openapi
+package openapigen
 
 import (
 	"errors"
